add validation

+++ src/team4/personaldietary/GUI/FxmlController.java
@@ -31,7 +31,14 @@ import team4.personaldietary.business.*;
 public class FxmlController {
     @FXML
     private BorderPane bPane;
-
+    TextField nameField;
+    TextField timeField;
+    // TextField groupField;
+    TextField servingField;
+    TextField mealField;
+    TextField typeField;
+    TextField retailerField;
+    ComboBox<String> foodGroup;
     private ObservableList<Dining> diningList = FXCollections.observableArrayList();
     private ListView<Dining> listView = new ListView<>(diningList);
     private DiningDAO diningDAO = new DiningDAOImpl();
@@ -68,29 +75,29 @@ public class FxmlController {
         // Choice box for meal
         gridPane.setAlignment(Pos.TOP_CENTER);
 
+        // here, it is better create a string array, then foreach loop generate text object
+        Text nameText = new Text("Name");
+        Text timeText = new Text("Time");
+        Text groupText = new Text("Group");
+        Text servingText = new Text("Serving");
+        Text mealText = new Text("Meal");
+        Text typeText = new Text("Type");
+        Text retailerText = new Text("Retailer");
 
-        Text text1 = new Text("Name");
-        Text text2 = new Text("Time");
-        Text text3 = new Text("Group");
-        Text text4 = new Text("Serving");
-        Text text5 = new Text("Meal");
-        Text text6 = new Text("Type");
-        Text text7 = new Text("Retailer");
-
-        TextField nameField = new TextField();
-        TextField timeField = new TextField();
+        nameField = new TextField();
+        timeField = new TextField();
         // TextField groupField = new TextField();
-        TextField servingField = new TextField();
-        TextField mealField = new TextField();
-        TextField typeField = new TextField();
-        TextField retailerField = new TextField();
+        servingField = new TextField();
+        mealField = new TextField();
+        typeField = new TextField();
+        retailerField = new TextField();
         // Since default indining/outdining toggle is indining, default gray out retailer field.
         retailerField.setDisable(true);
 
         Button addItemButton = new Button("Add Item");
 
         //Combo box for food group
-        ComboBox<String> foodGroup = new ComboBox<>();
+        foodGroup = new ComboBox<>();
         foodGroup.setPromptText("What group does it belong?");
         foodGroup.getItems().addAll("Vegetables and Fruits", "Grain Products", "Milk and Alternatives", "Meat and Alternatives");
 
@@ -163,27 +170,28 @@ public class FxmlController {
         EventHandler<ActionEvent> addEventHandler = new EventHandler<ActionEvent>() {
             @Override
             public void handle(ActionEvent actionEvent) {
-                Dining foodItem;
-                // if it is indining food item
-                if (inOutDining.get()) {
-                    foodItem = new Indining(nameField.getText(), timeField.getText(),
-                            stringToGroup(foodGroup.getValue()), servingField.getText(),
-                            mealField.getText(), typeField.getText());
-                }
-                // else, add outdining food item
-                else {
-                    foodItem = new Outdining(nameField.getText(), timeField.getText(),
-                            stringToGroup(foodGroup.getValue()), servingField.getText(),
-                            mealField.getText(), retailerField.getText());
+                Dining foodItem = null;
+                if(validateInput(nameField) && validateInput(timeField) && validateInput(servingField) && validateInput(mealField) && validateInput(foodGroup)){
+                    // if it is indining food item
+                    if (inOutDining.get() && validateInput(typeField)) {
+                        foodItem = new Indining(nameField.getText(), timeField.getText(),
+                                stringToGroup(foodGroup.getValue()), servingField.getText(),
+                                mealField.getText(), typeField.getText());
+                    }
+                    // else, add outdining food item
+                    else if(validateInput(retailerField)){
+                        foodItem = new Outdining(nameField.getText(), timeField.getText(),
+                                stringToGroup(foodGroup.getValue()), servingField.getText(),
+                                mealField.getText(), retailerField.getText());
+                    }
                 }
 
-                // call add food item function of the business layer
-                diningDAO.addDiningItem(diningList, foodItem);
-                markFoodGroupAdd(foodItem.getGroup());
+                if(foodItem != null) { // call add food item function of the business layer
+                    diningDAO.addDiningItem(diningList, foodItem);
+                    markFoodGroupAdd(foodItem.getGroup());
 
-                // clear the content in left panel after click the add button.
-                nameField.clear(); timeField.clear(); servingField.clear(); mealField.clear(); retailerField.clear();
-                typeField.clear(); foodGroup.getSelectionModel().clearSelection();
+                   refreshItems();
+                }
             }
         };
         addItemButton.setOnAction(addEventHandler);
@@ -196,19 +204,19 @@ public class FxmlController {
 
 
         // add item to gridPane
-        gridPane.add(text1, 0, 0);
+        gridPane.add(nameText, 0, 0);
         gridPane.add(nameField, 1, 0);
-        gridPane.add(text2, 0, 1);
+        gridPane.add(timeText, 0, 1);
         gridPane.add(timeField, 1, 1);
-        gridPane.add(text3, 0, 2);
+        gridPane.add(groupText, 0, 2);
         gridPane.add(foodGroup, 1, 2);
-        gridPane.add(text4, 0, 3);
+        gridPane.add(servingText, 0, 3);
         gridPane.add(servingField, 1, 3);
-        gridPane.add(text5, 0, 4);
+        gridPane.add(mealText, 0, 4);
         gridPane.add(mealField, 1, 4);
-        gridPane.add(text6, 0, 5);
+        gridPane.add(typeText, 0, 5);
         gridPane.add(typeField, 1, 5);
-        gridPane.add(text7, 0, 6);
+        gridPane.add(retailerText, 0, 6);
         gridPane.add(retailerField, 1, 6);
         gridPane.add(dining, 0, 7);
         gridPane.add(inOut, 1, 7);
@@ -236,12 +244,11 @@ public class FxmlController {
         removeButton.setOnAction(new EventHandler<ActionEvent>() {
             @Override
             public void handle(ActionEvent actionEvent) {
-                Dining foodItem = listView.getSelectionModel().getSelectedItem();
-
-                diningDAO.removeDiningItem(diningList, foodItem);
-
-                markFoodGroupRemove(foodItem.getGroup());
-
+                if(listView.getItems().size()>0) {
+                    Dining foodItem = listView.getSelectionModel().getSelectedItem();
+                    diningDAO.removeDiningItem(diningList, foodItem);
+                    markFoodGroupRemove(foodItem.getGroup());
+                }
             }
         });
 
@@ -260,12 +267,27 @@ public class FxmlController {
         bPane.setBottom(bottomMenu);
     }
 
+    private void refreshItems(){
+        // clear the content in left panel after click the add button.
+        nameField.clear();
+        timeField.clear();
+        servingField.clear();
+        mealField.clear();
+        retailerField.clear();
+        typeField.clear();
+        foodGroup.getSelectionModel().clearSelection();
+        foodGroup.setPromptText("What group does it belong?");
+        foodGroup.setStyle("-fx-background-color: LIGHTGRAY");
+    }
+
     private Group stringToGroup(String s) {
-        if (s.equals("Vegetables and Fruits")) return Group.vegetable_and_fruit;
-        else if (s.equals("Grain Products")) return Group.grain_products;
-        else if (s.equals("Milk and Alternatives")) return Group.milk_and_alternatives;
-        else if (s.equals("Meat and Alternatives")) return Group.meat_and_alternatives;
-        else return null;
+        switch (s){
+            case "Grain Products": return Group.grain_products;
+            case "Milk and Alternatives": return Group.milk_and_alternatives;
+            case "Meat and Alternatives": return Group.meat_and_alternatives;
+            case "Vegetables and Fruits": return Group.vegetable_and_fruit;
+            default: return null;
+        }
     }
 
     private void markFoodGroupAdd(Group group) {
@@ -304,5 +326,23 @@ public class FxmlController {
             numItemsInFoodGroup[3]--;
             if (numItemsInFoodGroup[3] <= 0) buttonMeat.setStyle("-fx-background-color: grey");
         }
+    }
+
+    private boolean validateInput(TextField textField){
+        if(textField.getText() == null ||textField.getText().isEmpty()){
+            textField.setPromptText("Invalid input");
+            textField.getParent().requestFocus();
+            return false;
+        }
+        return true;
+    }
+
+    private boolean validateInput(ComboBox<String> comboBox){
+        if(comboBox.getValue() == null ){
+            comboBox.setStyle("-fx-background-color: red");
+            comboBox.getParent().requestFocus();
+            return false;
+        }
+        return true;
     }
 }
